docs: renumbers slides
