+++ lesson_19/README.md
@@ -1,4 +1,4 @@
-# Lesson 22: Building Applications ([Slides](https://code-differently.github.io/code-society-25-2/slides/#/lesson_22))
+# Lesson 19: Building Applications ([Slides](https://code-differently.github.io/code-society-25-2/slides/#/lesson_19))
 
 ## Pre-work
